Testing + jar file

+++ src/main/java/servicii/foto/SedintaFotoProdus.java
@@ -19,15 +19,15 @@ public class SedintaFotoProdus extends ServiciiFoto {
     @Override
     public String informatiiServici() {
         if(album && !studio)
-            return super.informatiiServici() + ": Fotografie produs in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Fotografie produs in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face atat pe stick si veti avea si un album";
         else if(!album  && !studio)
-            return super.informatiiServici() + ": Fotografie produs in aer liber. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Fotografie produs in aer liber. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
         else if(!album)
-            return super.informatiiServici() + ": Fotografie produs in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+            return super.informatiiServici() + ": Fotografie produs in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                     + " $. Livrarea se va face pe stick";
-        else return super.informatiiServici() + ": Fotografie produs in studio. Sedinta va dura " + nrMinute + " minute si va costa " + (int)pretTotal
+        else return super.informatiiServici() + ": Fotografie produs in studio. Sedinta va dura " + (int)nrMinute + " minute si va costa " + (int)pretTotal
                 + " $. Livrarea se va face atat pe stick si veti avea si un album";
     }
 }
